+++ examples/boot_task.rs
@@ -1,8 +1,11 @@
 use planif::enums::TaskCreationFlags;
 use planif::schedule_builder::{Action, ScheduleBuilder};
+use planif::schedule::TaskScheduler;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let sb = ScheduleBuilder::new().unwrap();
+    let ts = TaskScheduler::new()?;
+    let com = ts.get_com();
+    let sb = ScheduleBuilder::new(&com).unwrap();
 
     sb.create_boot()
         .author("Matt")?
@@ -12,6 +15,5 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         .start_boundary("2022-04-28T02:14:08.660633427+00:00")?
         .build()?
         .register("TimeTaskName", TaskCreationFlags::CreateOrUpdate as i32)?;
-
     Ok(())
 }