+++ examples/monthly_task.rs
@@ -1,9 +1,13 @@
 use chrono::prelude::*;
 use planif::enums::{DayOfMonth, Month, TaskCreationFlags};
+use planif::schedule::TaskScheduler;
 use planif::schedule_builder::{Action, Monthly, ScheduleBuilder};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let builder: ScheduleBuilder<Monthly> = ScheduleBuilder::new()?
+    let ts = TaskScheduler::new()?;
+    let com = ts.get_com();
+
+    let builder: ScheduleBuilder<Monthly> = ScheduleBuilder::new(&com)?
         .create_monthly()
         .author("Matt")?
         .description("Test Trigger")?