+++ examples/various_settings_task.rs
@@ -2,9 +2,12 @@ use chrono::prelude::*;
 use planif::enums::TaskCreationFlags;
 use planif::schedule_builder::{Action, ScheduleBuilder};
 use planif::settings::Settings;
+use planif::schedule::TaskScheduler;
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let sb = ScheduleBuilder::new().unwrap();
+    let ts = TaskScheduler::new()?;
+    let com = ts.get_com();
+    let sb = ScheduleBuilder::new(&com).unwrap();
     let mut settings = Settings::new();
     settings.run_only_if_idle = Some(true);