+++ examples/weekly_task.rs
@@ -1,9 +1,12 @@
 use chrono::prelude::*;
-use planif::enums::{DayOfWeek, TaskCreationFlags};
+use planif::enums::{ DayOfWeek, TaskCreationFlags };
+use planif::schedule::TaskScheduler;
 use planif::schedule_builder::{Action, ScheduleBuilder};
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
-    let sb = ScheduleBuilder::new().unwrap();
+    let ts = TaskScheduler::new()?;
+    let com = ts.get_com();
+    let sb = ScheduleBuilder::new(&com).unwrap();
 
     sb.create_weekly()
         .author("Matt")?
@@ -15,6 +18,5 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         .weeks_interval(3)?
         .build()?
         .register("WeeklyTaskName", TaskCreationFlags::CreateOrUpdate as i32)?;
-
     Ok(())
 }